+++ lib/hexapdf/layout.rb
@@ -49,6 +49,7 @@ module HexaPDF
     autoload(:Frame, 'hexapdf/layout/frame')
     autoload(:WidthFromPolygon, 'hexapdf/layout/width_from_polygon')
     autoload(:TextBox, 'hexapdf/layout/text_box')
+    autoload(:ImageBox, 'hexapdf/layout/image_box')
 
   end